+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/implementation/servicebus/CustomAzureServiceBusConnectionDetails.java
@@ -1,0 +1,18 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.spring.cloud.autoconfigure.implementation.servicebus;
+
+import com.azure.spring.cloud.autoconfigure.implementation.servicebus.properties.AzureServiceBusConnectionDetails;
+
+import static com.azure.spring.cloud.autoconfigure.implementation.util.TestServiceBusUtils.CONNECTION_STRING_FORMAT;
+
+public class CustomAzureServiceBusConnectionDetails implements AzureServiceBusConnectionDetails {
+
+    static final String CONNECTION_STRING = String.format(CONNECTION_STRING_FORMAT, "connection-detail-namespace");
+
+    @Override
+    public String getConnectionString() {
+        return CONNECTION_STRING;
+    }
+}