+++ sdk/spring/spring-cloud-azure-testcontainers/src/main/java/com/azure/spring/cloud/testcontainers/implementation/service/connection/bus/ServiceBusContainerConnectionDetailsFactory.java
@@ -1,0 +1,34 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.spring.cloud.testcontainers.implementation.service.connection.bus;
+
+import com.azure.spring.cloud.autoconfigure.implementation.servicebus.properties.AzureServiceBusConnectionDetails;
+import org.springframework.boot.testcontainers.service.connection.ContainerConnectionDetailsFactory;
+import org.springframework.boot.testcontainers.service.connection.ContainerConnectionSource;
+import org.testcontainers.azure.ServiceBusEmulatorContainer;
+
+class ServiceBusContainerConnectionDetailsFactory
+    extends ContainerConnectionDetailsFactory<ServiceBusEmulatorContainer, AzureServiceBusConnectionDetails> {
+
+    @Override
+    protected AzureServiceBusConnectionDetails getContainerConnectionDetails(ContainerConnectionSource<ServiceBusEmulatorContainer> source) {
+        return new ServiceBusContainerConnectionDetails(source);
+    }
+
+    /**
+     * {@link AzureServiceBusConnectionDetails} backed by a {@link ContainerConnectionSource}.
+     */
+    private static class ServiceBusContainerConnectionDetails extends ContainerConnectionDetails<ServiceBusEmulatorContainer>
+        implements AzureServiceBusConnectionDetails {
+
+        ServiceBusContainerConnectionDetails(ContainerConnectionSource<ServiceBusEmulatorContainer> source) {
+            super(source);
+        }
+
+        @Override
+        public String getConnectionString() {
+            return getContainer().getConnectionString();
+        }
+    }
+}